+++ full-node/sov-ledger-apis/src/lib.rs
@@ -1,5 +1,6 @@
 #![forbid(unsafe_code)]
 
+#[cfg(feature = "server")]
 pub mod rest;
 pub mod rpc;