+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- #725 removes the `macros` feature from `sov-modules-api`, which is now always enabled even with `--no-default-features`.
 - #714 integrates a batch of changes to the `StfBlueprint` and the capabilities. Meaningful changes:
   - Remove arguments of type `SequencerStakeMeter` from the capabilities and replace all of the `(SequencerStakeMeter, StateCheckpoint)` couple of variables by a single `PreExecWorkingSet` which is a type safe data structure that should charge for gas before transaction execution starts.
   - Removing the `ExecutionMode` type from the `StfBlueprint`. It is now replaced by `TxScratchpad`, which is an intermediary type between `WorkingSet` and `StateCheckpoint`. This is useful for the `sov-sequencer` crate because one may want to revert all the changes that happened during transaction execution when simulating a transaction before adding it to a sequencer batch.